moved macro functions to __private subpackage with warning
